Add 'by' clauses to distinguish resource kinds in the calculation
To handle cases where custom flow schemas are created to throttle
only certain kind of resources, the ratio used in the alert expression
has to be calculated using the terminated and total requests relevant
to those resource kinds.

Without accounting for the resource kind, it's possible that the alert
won't fire because the global apiserver_request_total is many magnitude
higher than the apiserver_request_terminations_total, even though the
number of terminated requests for certain resource kinds are not
negligible due to the kind of custom flow schemas described above.

Signed-off-by: Ivan Sim <isim@redhat.com>